WIP. `joinScripts` partially propagated
Does not build (yet)

Looks like this change has been anticipated for a while (years?).
I have some thoughts to discuss before pursuing it.

+ "ScriptReference/ ReferencedScript/ Inline Script".
These names are not too absurd. And partially aligns with the vocab
used in the docs/spec when this was fudged in.
It is confusing though. It should be "NoScript/Script" and elsewhere "ScriptHash".

+ `Input` can accommodate both `refScriptHash` and `refScript`, but the
  latter includes the former.
  I think this can be flattened.